my-flask-app1/requirements.txt
flask
azure-functions

